feat: support parsing and generating d.ts for @event
This change is very specific to WebViewer and cannot be applied for JSDoc in other programs.
When a @event appears, the associated class in WebViewer extends the EventHandler, which means the class has methods like `on`, `off`, `one` and `trigger`. This PR makes the change so that whenever a @event tag is visited, three TypeScript method declarations(on, off, one) will be added to the root tree, and which will be emitted later